129: split services/runs.rs by concern (PR-G4)
Splits the 1197-line services/runs.rs (577 non-test) into two files
under services/runs/, both well under rust.md § 13's ~500-line
non-test threshold:

- runs/submission.rs (~300 non-test): create_run with its validation
  pipeline (validate_time_fields, validate_run_request, insert_run) and
  delete_run. Plus the CreateRunRequest input DTO and the private
  ValidatedRunTimes shape.
- runs/read.rs (~250 non-test): get_run / list_runs / get_run_defaults
  with their response DTOs (RunDetail, RunDefaults, RunFilters) and the
  private RunDetailRow row-shape.

The new services/runs.rs is a thin module root that declares the two
submodules and re-exports their public items, so callers continue to
reach everything via crate::services::runs::<name>.

Tests follow their production code into each submodule's
#[cfg(test)] mod tests. Pure validation tests + create/delete tests +
ordered-submit-guard tests live in submission.rs; list/defaults tests
live in read.rs.

git mv was used to move the original runs.rs to runs/submission.rs (the
biggest extract at ~70% content similarity). Because the new runs.rs
module root reuses the original path, default `git log --follow`
(which doesn't enable break-rewrites) sees runs.rs as modified rather
than deleted, so it won't trace through to submission.rs at default
settings. Use `git log --follow -B -M30 -C` or `git blame -C -C` for
function-level history across the split.

Per docs/compliance-plan.md PR-G4 / Issue #129. No API or schema change.
All 36 affected unit tests (now distributed across submission.rs and
read.rs) plus the existing integration tests pass unchanged.
